requirements.txt
mysql-connector-python
bcrypt
pyperclip
